Fundamental Principle of Monte Carlo Simulation

\[p_{\Re} (\Re) \equiv \begin{cases} 1 & \quad 0 \leq \Re \leq 1, \\ 0 & \quad \Re < 0 \cup \Re < 1. \end{cases}\]

The Optical Path Length

\[r=-\frac{1}{c}\ln{\Re}\]

Sampling Scattering direction

azimuthal angle in the plane of the scattering event relative to the direction of photons before scattering $\varphi$ is uniformly distribute between 0 and $2\pi$ Therefore,

\[\varphi = 2\pi\Re\]

To find the angle between new trajectory and the direction of photons before scattering, we use the Petzold..

LightMC.phasePetzold — Method

phasePetzold()

return 2 arrays: ϕps and θps. When ϕps is the cumulation distribution of scattering angle and θps is the angle between new trajectory and the direction of the photon before scattering corresponding to each ϕps For example, the probability to find the scattering photon from the direction of the photon before scattering to θps[1] is ϕps[1] The data come from Kirk,1981, Monte Carlo Procedure for Simulating the Penetration of Light into Natural Waters